Tweak series titles and synopses
I think removing the part numbers makes them a bit more interesting on
their own, without creating confusion about the true title of the post
by having a different title on the page than in the metadata and
overview page. We do omit the part number from the plain text title,
but I think that is fine.