Make RCL examples compatible with 0.5.0
Should I go back and edit previously published blog posts to update them
for the current version of a tool, or should I leave the historical
record to make it clear how things evolved? Generally I lean towards the
latter, but in this case I think the post may be a significant learning
material; the changes are superficial; and the full history is still
public in the Git repository, I'm not actively trying to hide these
edits. So I think in this case it's acceptable.