Explore the design space
I am happy with this phrasing, it makes the trade-off very explicit, and
we can find examples of each category. It also sets RCL apart from other
languages, and doing something else may be interesting to explore. Or
maybe I am now being a fool to waste the innovation tokens on making
numbers work in a weird way, when so many other languages already have
numbers that seem to work just fine.