Tweak conclusions, delete table
There is no alignment chart any more. It's a shame, it would have been
a nice way to present things if the structure of the problem admitted
that, but I don't think that it does any more. It was a valuable tool
to explore the design space though. The current set of conflicting
wishlist items was not obvious to me when I started all this! Writing up
this post helped to clarify all this.