Remove comma from date format
According to Wikipedia, in UK English, one does not use a comma in the
date format (and also not ordinals, which I was already not using).
It is used in US English, because you have two numbers next to each
other. I switched from UK to US English for posts at some point, but
the date format, YYYY-MM-DD is still the best of course, but if we
_have_ to do something in prose, dd mmm yyyy makes far more sense than
the insane US format.