Add nav header at the top of the page
This has long been a problem on my blog, I don't really have a heading
where people can navigate back to the index. The author name links back
home, but that is not obvious. So a subtle navigation menu is in order,
but how to it without disturbing the clean article lay-out? I think this
dimmed one works, even when the contrast is a bit low. It is supposed to
be subtle.