Include nav header on all pages, extend navigation
Adding the header everywhere ensures the title is in the same place when
navigating between pages. Even if the nav is duplicated, I think that's
okay. Add Codeberg to the nav everywhere.

Also make the page a bit wider so my name does not line-wrap in the wide
case.

Also, since my new posts are in US English, rephrase the index and
footer in US English. Would be neat to make the phrase in the footer
adjust to the language of the post.