Finish first version of Musium intro
I need to add the image to the blog here because of course I made the
generator peek into the metadata, I can't just hotlink the one that the
same server is already serving for docs.ruuda.nl. But that's fine. Maybe
I should also compress it with webp or something.